设置app.vue登陆
