+++ src/page/modules/axios/index.vue
@@ -5,10 +5,8 @@
 </template>
 
 <script>
-import moment from 'moment'
-import { set } from '../../../utils/cookie.js'
 export default {
-  name: '',
+  name: 'axios',
 
   data() {
     return {}
@@ -33,30 +31,10 @@ export default {
           userid: 3462
         }
       })
-    },
-    login() {
-      this.$callApi({
-        api: 'user_user_login',
-        param: {
-          loginName: 'dddddddd',
-          password: 'zaq1234'
-        }
-      }).then(data => {
-        set({
-          key: 'JSSSID_COOKIE',
-          value: data.token,
-          domain: document.location.hostname,
-          expires: moment()
-            .add(2, 'hours')
-            .toDate()
-            .toUTCString()
-        })
-        this.reFindCount()
-      })
     }
   },
   created() {
-    this.login()
+    this.reFindCount()
   }
 }
 </script>
